requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
asyncpg==0.30.0
redis==5.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic==2.10.4
pydantic-settings==2.7.1
email-validator==2.2.0
httpx==0.28.1
openai==1.58.1
anthropic==0.42.0
google-generativeai==0.8.3
pinecone-client==5.0.1
pypdf==5.1.0
python-docx==1.1.2
python-multipart==0.0.20
pandas==2.2.3
celery==5.4.0
alembic==1.14.0
structlog==24.4.0
slowapi==0.1.9
pytest==8.3.4
pytest-asyncio==0.25.0
